+++ server/lib/nl/common/utils.py
@@ -26,7 +26,7 @@ import server.lib.nl.common.counters as ctr
 import server.lib.nl.common.utterance as nl_uttr
 from server.lib.nl.detection.types import ClassificationType
 import server.lib.nl.detection.types as types
-import server.lib.nl.fulfillment.context as ctx
+import server.lib.nl.fulfillment.utils as futils
 import shared.lib.constants as shared_constants
 
 # TODO: Consider tweaking/reducing this
@@ -225,10 +225,10 @@ def parent_place_names(dcid: str) -> List[str]:
 def get_contained_in_type(
     uttr_or_classifications: Any) -> types.ContainedInPlaceType:
   if isinstance(uttr_or_classifications, nl_uttr.Utterance):
-    classification = ctx.classifications_of_type_from_utterance(
+    classification = futils.classifications_of_type_from_utterance(
         uttr_or_classifications, types.ClassificationType.CONTAINED_IN)
   else:
-    classification = ctx.classifications_of_type(
+    classification = futils.classifications_of_type(
         uttr_or_classifications, types.ClassificationType.CONTAINED_IN)
   place_type = None
   if (classification and isinstance(classification[0].attributes,
@@ -239,7 +239,7 @@ def get_contained_in_type(
 
 
 def get_size_types(uttr: nl_uttr.Utterance) -> List[types.SizeType]:
-  classification = ctx.classifications_of_type_from_utterance(
+  classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.SIZE_TYPE)
   size_types = []
   if (classification and isinstance(classification[0].attributes,
@@ -250,7 +250,7 @@ def get_size_types(uttr: nl_uttr.Utterance) -> List[types.SizeType]:
 
 
 def get_ranking_types(uttr: nl_uttr.Utterance) -> List[types.RankingType]:
-  classification = ctx.classifications_of_type_from_utterance(
+  classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.RANKING)
   ranking_types = []
   if (classification and isinstance(classification[0].attributes,
@@ -262,7 +262,7 @@ def get_ranking_types(uttr: nl_uttr.Utterance) -> List[types.RankingType]:
 
 def get_quantity(
     uttr: nl_uttr.Utterance) -> types.QuantityClassificationAttributes:
-  classification = ctx.classifications_of_type_from_utterance(
+  classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.QUANTITY)
   if (classification and isinstance(classification[0].attributes,
                                     types.QuantityClassificationAttributes)):
@@ -271,7 +271,7 @@ def get_quantity(
 
 
 def get_time_delta_types(uttr: nl_uttr.Utterance) -> List[types.TimeDeltaType]:
-  classification = ctx.classifications_of_type_from_utterance(
+  classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.TIME_DELTA)
   time_delta = []
   # Get time delta type
