+++ server/lib/nl/fulfillment/event.py
@@ -18,7 +18,7 @@ from server.lib.nl.common import utils
 import server.lib.nl.common.utterance as nl_uttr
 from server.lib.nl.detection import types
 import server.lib.nl.fulfillment.base as base
-import server.lib.nl.fulfillment.context as ctx
+import server.lib.nl.fulfillment.utils as futils
 
 _DEFAULT_EVENT_PLACE = types.Place("country/USA", "USA", "Country")
 
@@ -28,7 +28,7 @@ _DEFAULT_EVENT_PLACE = types.Place("country/USA", "USA", "Country")
 
 
 def populate(uttr: nl_uttr.Utterance) -> bool:
-  event_classification = ctx.classifications_of_type_from_utterance(
+  event_classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.EVENT)
   if (not event_classification or
       not isinstance(event_classification[0].attributes,
@@ -38,7 +38,7 @@ def populate(uttr: nl_uttr.Utterance) -> bool:
     return False
   event_types = event_classification[0].attributes.event_types
 
-  ranking_classification = ctx.classifications_of_type_from_utterance(
+  ranking_classification = futils.classifications_of_type_from_utterance(
       uttr, types.ClassificationType.RANKING)
   ranking_types = []
   if (ranking_classification and
